Add gdb enable and port flag to ops run
Added support for enabling qemu's GDB interface. Only debugging over TCP
is supported (fortunately, it's the most common option, and the most
useful for us), and the user can specify the TCP port to be used for the
gdb interface. Specifying a port also enables the gdb interface --
specifying one without the other would be meaningless, since only TCP is
supported.

Signed-off-by: Alexandru Lazar <alazar@startmail.com>