add attach volume cmd and mount on run
`ops volume attach` lets user attach volume to a stopped instance.
as of this commit, `ops run` overwrite this config though.

`ops run --mounts` lets user mount volume to an instance on run, but
does not persist the config. as of this commit, changes made to volume
during run is visible, but gets discarded on subsequent run without
`--skipbuild`