ldd_linux: fix check for dynamically linked executable
The presence of a dynamic section in an ELF file does not
necessarily make the file dynamically linked: a file may have a
dynamic section while not having an dependency on shared libraries.
The ArangoDB server executable binary for Linux (/usr/sbin/arangod
file in https://download.arangodb.com/arangodb37/Community/Linux/arangodb3_3.7.6-1_amd64.deb)
is an example of such files.
When a statically linked executable is run with
LD_TRACE_LOADED_OBJECTS=1 (as done in the getSharedLibs() method),
the dynamic linker executes the binary instead of listing its
shared libraries, and this makes getSharedLibs() fail; as a result,
"ops run" does not work.

An ELF file is dynamically linked only if it has a dynamic section
and this section contains at least one DT_NEEDED entry. The
ImportedLibraries() method from the debug/elf Go package can be used
to retrieve the list of shared libraries.
This commit changes IsDynamicLinked() to call ImportedLibraries()
instead of looking for the dynamic section in the ELF file.