Hyper-V/Azure: add support for UEFI-based Generation-2 instances
When creating an image for Hyper-V or Azure, the `--type` command
line flag (or alternatively, the "ImageType" attribute under
"CloudConfig" in the configuration file) can be used to specify a
Generation-2 image type, that can run on Generation-2 instances
(as well as Generation-1 instances, except for some types of
Azure VMs).
In Azure, the VM generation when creating an instance is chosen
automatically based on the image type; the default instance flavor
has been changed from A1V2 to B1s, because A1V2 instances don't
support Generation-2 images.
To specify the VM generation when creating a Hyper-V instance, the
`-f` command line flag (or alternatively, the Flavor attribute
under "CloudConfig" in the configuration file) can be used.
Configuration file example for Hyper-V:
```
{
    "CloudConfig": {
        "ImageType":"gen2",
        "Flavor": "gen2"
    }
}
```
Images created with the "gen2" option have an additional partition
containing the UEFI bootloader file (which will be part of future
Nanos releases).