Qemu: add support for running non-raw Linux images
Debian Linux images cannot be run by Qemu with the `-vga none`
option; in addition, when the image file has a format other than
raw, Qemu must not be passed the `format=raw` option and should be
allowed to automatically detect the file format.
This will be needed in the next commit to add support for
cross-build environments.