Provider CreateVolume(): simplify processing of size option
The size of a volume being created can be supplied by the user via
either the --size command line option or the BaseVolumeSz
configuration option. This commit modifies the volume create
command handler so that any size value supplied by the user is
stored in BaseVolumeSz, and the command line option value is not
passed separately to the provider implementation of CreateVolume().
This allows CreateVolume() implementations to process the volume
size without having to check whether the size value has been
supplied as a command line option or as a configuration option.
To convert a size value in GB, the GetSizeInGb() helper function
has been created.