AWS volume create: avoid uploading large files to S3
The AWS volume create implementation first uploads to S3 a locally
created volume, then creates a snapshot from the uploaded file, and
finally creates a volume from the shapshot. When a volume is
created from a snapshot, it is possible to specify a size parameter
(which must be larger than or equal to the size of the snapshot):
this allows to create arbitrarily sized volumes from small
snapshots.
This commit changes the AWS volume create implementation so that
the volume file being uploaded to S3 is created with the minimum
size needed for the filesystem, and any volume size specified by
the user is passed to the AWS volume creation API command.