AWS volumes: dynamically generate device name when attaching volume
Using a single hard-coded device name "/dev/sdf" for all volume
attachments doesn't allow to attach more than one volume to a given
instance: when trying to attach a second volume, the result is
"InvalidParameterValue: Invalid value '/dev/sdf' for unixDevice.
Attachment point /dev/sdf is already in use".
This change replaces the hard-coded device name with a name that is
dynamically generated based on the existing block device mappings
on a given instance. In addition, it removes the device name from
the volume detach command, because there is no need to specify the
attachment point if both the volume ID and the instance ID are
specified.