run command: add option to specify RAM size
With the '-m' (or '--memory') command line flag, the 'ops run'
command assigns the specified amount of RAM to the local instance
being run. The flag value uses the same syntax as the
"RunConfig.Memory" attribute in the configuration file (e.g. "64M",
or "4G"), and if present overrides the value in the configuration
file.