Volumes: do not look for local volumes when using a cloud provider
When building an image containing volume mount directives, local
volumes are involved only in case of on-prem images, and not when
operating with a cloud provider. This commit removes from the
AddMounts() function the check for the presence of a local volume,
and moves this function from the onprem package to the lepton
package, because it is not specific to on-prem deployments; in
addition, the AddMountsFromConfig() function (which deals with
local volumes) is only called when starting an on-prem instance.