Handle read-only flag appended to mount paths
The mount path for a volume may now optionally have ":ro" appended to it
indicating to mount read-only. AddMount must be changed to handle trimming
this from the path in order to create a correct mount directory and
AddMounts changed to save the flag back to the mount path after
removing it.