GCP: add support for associating a service account to an instance
When executing the `ops instance create` command to create a GCP
instance, if the CloudConfig.InstanceProfile configuration
parameter is a non-empty string, the instance being created is
associated to the service account identified by this string, with
`cloud-platform` access scope. GCP service accounts are identified
by an email address; the special string "default" indicates the
default service account for a given project.
The service account specified in the configuration must exist in
the GCP project when an instance is being created, otherwise an
error is returned.
For more information about GCP service accounts, see
https://cloud.google.com/compute/docs/access/service-accounts.