On premises instance: add support for VirtFS (9P filesystem) (#1430)
This change adds support for exporting host folders to the guest
when running a local instance. Handling of the mount configuration
options is being amended so that if there are no local volumes
matching a label in a mount directive, a folder in the host
filesystem is exported to the guest via VirtFS instead of using a
pre-made volume.
This new functionality can be used in the 'run' and 'pkg load'
commands, both as a command line option (e.g.
'--mounts host_dir:/guest_dir') and as a JSON attribute in the
configuration file.