Image creation: add support for images without a bootloader
With this change, if the `Boot` attribute in the configuration is
empty, Ops does not require a boot.img file: if such file is found,
it is inserted in the image being created, otherwise Ops builds
internally a partition table when creating the image.
This allows crating images without a legacy bootloader (e.g. images
with a UEFI bootloader, or images with no bootloader at all, which
can be used when the kernel binary file is passed directly to the
hypervisor).
As part of these changes, if the user specifies in the
configuration a kernel and/or boot file path, these settings are no
longer overridden by the nightly flag.