Proxmox instance: add serial port device (#1495)
This allows using the xterm.js console tool from the Proxmox web
interface to see the output messages printed by an instance.
xterm.js is a terminal emulator that captures the serial port logs
output by an instance, and is not affected if the instance resets
itself, thus allowing users to see e.g. crash messages if an
instance reboot is triggered by a crash.
The alternative tool for seeing instance output via the Proxmox web
intercace is noVNC, which captures the logs sent to the VGA device;
this tool is still available, and can be used independently from
xterm.js.