re-add support for TFS version 4 encodings
This commit returns support for reading and optionally writing TFS version 4
images. 'ops image' commands that read an existing image will read either
version 4 or 5 images, and commands that create images may either accept a
"-4" or "--tfsv4" commandline option or a "TFSv4" config bool to create a
version 4 image.

This support is intended to help ease the transition towards use of TFS
version 5.