AWS images: set ENA support to true regardless of instance type
The EnaSupport flag of the ec2.RegisterImageInput structure
indicates the the OS in an image has an ENA driver, not that it
requires an ENA device; thus, an image flagged as ENA-capable can
run just fine on AWS instances without an ENA adapter (such as
non-Nitro instance types).
This change removes the NitroInstanceTypes map and the
GetEnaSupportForFlavor() function, and sets the EnaSupport flag to
true regardless of the instance type. This allows creating images
that can run on newer Nitro instance types such as m7a, m7i and
r7iz.