AWS: add support for Graviton3 (ARM) instances
This allows creating images to be run on AWS m7g and r7g instances.
The list of Graviton-based instances is being removed since it is
not needed to determine whether UEFI should be enabled (all
ARM-based instances require a UEFI image).
In addition, UEFI enablement is being added when creating an image
from a package.